Post-Duluth Autonomous Command Cleanup Beginnings 21:46
Didn't get all the way through, but some standards were set alongside
the normal cleanup.
Trajectories start with t_
Ramsetes start with r_
Trajectories and ramsetes use the same name besides the t_/r_ prefix.
Each trajectory/ramsete name will describe what its purpose is.
Each trajectory will have a general idea of steps to be completed once the
trajectory endpoint is reached. (Up for debate on whether this is the
place for it or not)